Create README.md
A handy script for sorting photos and videos
This simple script allows you to sort photos and video files. 
In the script, specify the tracking folder and specify the address of the folder for sorting photos and videos. 
In the second folder, create a new folder "Photos" and "Videos". 
Then take off the photos and videos of the first folder, they are automatically sorted by the second folder.